Declarations with optionals and defaults
WDL assumes all declarations to a task (also workflow) can be assigned by the caller. However, there are lots of declarations that are, in essence, temporary variables that hold calculations. In order to compile a task to an applet, we limit the applet inputs to unassigned expressions.

For example, `x` is an input, `y` is not. Is `z` an input?

```
Int x 
Int y = 6
Int? z = 17
```

We want to be able to have z as an input as well. This supports inputs with defaults.