Handle null repo (#437)
TaskRunner fails if RepoTags is missing or has a null value in the manifest of the docker tarball. However, docker load prints out the repo name. This change uses the results of docker load if the attempt to read the repo if reading from the manifest fails.

There's also some code reformatting here due to my scalafmt git hook. The only real changes are in TaskRunner and TaskRunnerTest.