Break up large scatters (#444)
This is a change to the runtime for scatter jobs. Instead of launching all subjobs at once, only a maximum of N subjobs (where N defaults to 500 but is user configurable between 1-1000) are launched at once. When one chunk finishes, the next launches, and so on until all sub-jobs have been executed.