Fix name regexp in DxFindDataObjects (#446)
In DxFindDataObjects, if a set of object names is specified, the query includes a regular expression search. The regular expression is incorrect. For example, if there are a set of app/workflow names {"foo", "bar", "baz"}, the code in dxWDL builds a regular expression to match any of them:

[(foo)|(bar)|(baz)]

which will match any of the characters f, o, b, a, r, z

When searching for objects in a folder with existing objects, the query takes extremely long to run. Most of the time the user thinks the program has hung and gives up (e.g. https://jira.internal.dnanexus.com/browse/APPS-140).

The correct regexp is ^foo$|^bar$|^baz$